auto-improve: Merge triplicated `_pr_fixture` test helper

Refs damien-robotsix/robotsix-cai#1319
